Fix Plant watering scheduling with past StartDate
When a Plant's StartDate is in the past, the scheduling will not be
calculated as I expected. I opened an issue with goCron but wanted to
implement a temporary fix for if that change is not desired.

Also added a small change to set the start date if it does not exist
when reading from YAML file.