Move "skip watering" feature to Go program
Now that watering is mostly controlled by the Go program, I removed
skipping from the microcontroller. Now I can configure the amount of
waterings to skip.
I removed the skip action and just made it part of updating a Plant's
fields instead. Now, watering events decrement the count and the
Plant has to be saved to the storage backend again