Fix backwards-compatible Plant action
Since the backwards-compatible middleware creates a "fake" Garden
containing all Plants, it doesn't have a valid MQTT topic. To fix this,
I just add a check to the action for this specific case. I might
revisit this to come up with a better solution that is fully handled
in the middleware